+++ src/api/yandex/__init__.py
@@ -1,7 +1,8 @@
 from .methods import (
     get_access_token,
     upload_file_with_url,
-    create_folder
+    create_folder,
+    publish
 )
 from .requests import (
     make_link_request,