+++ src/blueprints/telegram_bot/webhook/commands/help.py
@@ -50,6 +50,10 @@ def handle():
         "Original name will be saved. "
         "You can send direct URL to a file without this command."
         "\n"
+        f"{CommandsNames.PUBLISH.value} — publish a file or folder that already exists. "
+        "Send full name of item to publish with this command. "
+        f'Example: {to_code(f"/{yd_upload_default_folder}/files/photo.jpeg")}'
+        "\n"
         f"{CommandsNames.CREATE_FOLDER.value}— create a folder. "
         "Send folder name to create with this command. "
         "Folder name should starts from root, "